bot/requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
flask==2.3.2
flask-cors==3.0.10
